feat: 16 bit instructions

+++ include/vm.h
@@ -21,7 +21,7 @@ typedef enum {
  */
 typedef struct {
 	ObjectClosure *closure;
-	uint8_t *ip;
+	uint16_t *ip;
 	Value *slots;
 } CallFrame;
